refactor: extract Watcher config to main.rs
Also, rename `notifs` module to `notifications`
