+++ src/lib/conf.rs
@@ -1,20 +1,12 @@
 use std::collections::HashMap;
-use std::fs::{create_dir_all, File};
-use std::io::{Read, Write};
 
-use anyhow::Result;
-use glob;
-use log::{debug, error, trace, warn};
 use serde::{Deserialize, Serialize};
 
-use crate::notifs::Notifier;
-use crate::utils::cache_dir;
 use crate::utils::docker::{rng_docker_chars, DOCKER_NAME_CHARSET};
-use crate::utils::git::fetch;
 
 #[cfg(test)]
 mod tests {
-    use crate::conf::{FakeCIBinaryConfig, Image};
+    use crate::conf::Image;
     use crate::utils::tests::{deser_yaml, get_sample_resource_file};
 
     #[test]
@@ -51,24 +43,6 @@ mod tests {
                 panic!("got invalid image variant: {:?}", s);
             }
         }
-    }
-
-    #[test]
-    fn notifier_config() {
-        let c = get_sample_resource_file("notifiers.yml").expect("not found");
-        let conf: FakeCIBinaryConfig = serde_yaml::from_str(&c).expect("Could not parse yaml");
-        assert_eq!(conf.repositories.len(), 1);
-        let _: () = conf
-            .repositories
-            .iter()
-            .map(|repo| {
-                if let Some(notifiers) = &repo.notifiers {
-                    assert_eq!(notifiers.len(), 1);
-                } else {
-                    panic!("notifiers was None");
-                }
-            })
-            .collect();
     }
 }
 
@@ -156,133 +130,4 @@ impl Image {
             Image::Build(b) => b.name.clone(),
         }
     }
-}
-
-#[derive(Serialize, Deserialize, Debug, Eq, PartialEq)]
-#[serde(untagged)]
-pub enum BranchesSpec {
-    Single(String),
-    Multiple(Vec<String>),
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct FakeCIBinaryRepositoryConfig {
-    pub name: String,
-    pub uri: String,
-    pub branches: BranchesSpec,
-    pub notifiers: Option<Vec<Notifier>>,
-    #[serde(skip, default)]
-    pub refs: HashMap<String, String>,
-    #[serde(skip, default)]
-    pub br_regexps: Vec<glob::Pattern>,
-}
-
-impl FakeCIBinaryRepositoryConfig {
-    // horribly inefficient function.
-    // Hopefully we won't meet a repo with millions of branches.
-    pub fn update_branches(&mut self) -> Result<HashMap<String, String>> {
-        let mut diff = HashMap::new();
-        let r = fetch(&self.uri)?;
-        let deleted: Vec<String> = self
-            .refs
-            .keys()
-            .filter(|k| !r.contains_key(*k))
-            .map(|k| k.to_string())
-            .collect();
-        for d in &deleted {
-            self.refs.remove(d);
-        }
-        let mut changed: HashMap<String, String> = HashMap::from_iter(
-            r.iter()
-                .filter(|(k, _)| !self.refs.contains_key(*k))
-                .map(|(k, v)| (k.to_string(), v.to_string())),
-        );
-        diff.extend(changed.iter().map(|(k, v)| (k.to_string(), v.to_string())));
-        for (k, v) in self.refs.iter() {
-            if r.contains_key(k) && r.get(k).unwrap() != v {
-                changed.insert(k.to_string(), r.get(k).unwrap().to_string());
-                diff.insert(k.to_string(), r.get(k).unwrap().to_string());
-            }
-        }
-        self.refs.extend(changed);
-        Ok(diff)
-    }
-
-    pub fn init(&mut self) {
-        let v = match &self.branches {
-            BranchesSpec::Single(s) => {
-                trace!("Compiling branch pattern {}", s);
-                vec![glob::Pattern::new(s)
-                    .unwrap_or_else(|_| panic!("could not compile regex {}", s))]
-            }
-            BranchesSpec::Multiple(v) => v
-                .iter()
-                .map(|s| {
-                    trace!("Compiling branch pattern {}", s);
-                    glob::Pattern::new(s)
-                        .unwrap_or_else(|_| panic!("could not compile regex {}", s))
-                })
-                .collect(),
-        };
-        self.br_regexps = v;
-        // find cache dir
-        let cache = cache_dir();
-        // read cache dir
-        let mut s = String::new();
-        let fname = cache.join(format!("{}.yml", self.name));
-        let mut f = match File::open(&fname) {
-            Ok(f) => f,
-            Err(_e) => {
-                warn!(
-                    "Could not open file {} for persisted branch info",
-                    fname.display()
-                );
-                return;
-            }
-        };
-        let _ = f.read_to_string(&mut s);
-        let refs: HashMap<String, String> = match serde_yaml::from_str(&s) {
-            Ok(h) => h,
-            Err(_) => {
-                error!("could not deserialize file cache content, using fresh values");
-                return;
-            }
-        };
-        self.refs.extend(refs);
-    }
-
-    pub fn persist(&self) -> Result<()> {
-        trace!("persist()");
-        // find cache dir
-        let cache = cache_dir();
-        trace!("cache: {}", cache.display());
-        create_dir_all(&cache)?;
-        let mut f = File::create(cache.join(format!("{}.yml", self.name)))?;
-        // write to cache dir
-        let _ = f.write_all(serde_yaml::to_string(&self.refs)?.as_ref());
-        debug!("Finished persisting branch values to disk");
-        Ok(())
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-/// Config for the binary
-/// ```
-/// use fakeci::conf::FakeCIBinaryConfig;
-/// let s: &str = "repositories:
-///   - name: blabla
-///     uri: https://github.com/paulollivier/fake-ci
-///     branches: \"*\"";
-/// let c: FakeCIBinaryConfig = serde_yaml::from_str(s).expect("invalid yaml");
-/// assert_eq!(c.watch_interval, 300);
-/// assert_eq!(c.repositories.len(), 1);
-/// ```
-pub struct FakeCIBinaryConfig {
-    #[serde(default = "watch_interval_default")]
-    pub watch_interval: u32,
-    pub repositories: Vec<FakeCIBinaryRepositoryConfig>,
-}
-
-fn watch_interval_default() -> u32 {
-    300
 }
